More code improvements

+++ src/main/java/com/bbn/hadder/listener/LinkListener.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 
 public class LinkListener extends ListenerAdapter {
 
-    Rethink rethink;
+    private Rethink rethink;
 
     public LinkListener(Rethink rethink) {
         this.rethink = rethink;
@@ -26,44 +26,38 @@ public class LinkListener extends ListenerAdapter {
         event.getChannel().retrieveMessageById(event.getMessageId()).queue(
                 msg -> {
                     if (event.getMember().hasPermission(Permission.MANAGE_CHANNEL)) {
-                        if (!event.getMember().getUser().equals(event.getJDA().getSelfUser())) {
-                            if (msg.getAuthor().equals(event.getJDA().getSelfUser())) {
-                                if (msg.getEmbeds().size() == 1) {
-                                    if (msg.getEmbeds().get(0).getFooter() == null) {
-                                        if (msg.getEmbeds().get(0).getTitle().endsWith(") wants to link guilds!")) {
-                                            String requestguild = msg.getEmbeds().get(0).getTitle().replaceAll("\\) wants to link guilds!", "");
-                                            String requestguildid = null;
-                                            for (int i = requestguild.length() - 1; i >= 0; i--) {
-                                                if (String.valueOf(requestguild.charAt(i)).equals("(")) {
-                                                    requestguildid = requestguild.substring(i + 1);
-                                                    break;
-                                                }
-                                            }
-                                            if (requestguildid != null) {
-                                                if (event.getReactionEmote().getName().equals("✅")) {
-                                                    rethink.addLinkedGuild(event.getGuild().getId(), requestguildid);
-                                                    rethink.addLinkedGuild(requestguildid, event.getGuild().getId());
+                        if (!event.getMember().getUser().equals(event.getJDA().getSelfUser()) && msg.getAuthor().equals(event.getJDA().getSelfUser()) && msg.getEmbeds().size() == 1 && msg.getEmbeds().get(0).getFooter() == null && msg.getEmbeds().get(0).getTitle().endsWith(") wants to link guilds!")) {
+                            String requestguild = msg.getEmbeds().get(0).getTitle().replaceAll("\\) wants to link guilds!", "");
+                            String requestguildid = null;
+                            for (int i = requestguild.length() - 1; i >= 0; i--) {
+                                if (String.valueOf(requestguild.charAt(i)).equals("(")) {
+                                    requestguildid = requestguild.substring(i + 1);
+                                    break;
+                                }
+                            }
+                            if (requestguildid != null) {
+                                if (event.getReactionEmote().getName().equals("✅")) {
+                                    rethink.addLinkedGuild(event.getGuild().getId(), requestguildid);
+                                    rethink.addLinkedGuild(requestguildid, event.getGuild().getId());
 
-                                                    msg.delete().queue();
+                                    msg.delete().queue();
 
-                                                    MessageEmbed msgembed = new MessageEditor().setDefaultSettings(MessageEditor.MessageType.INFO)
-                                                            .setTitle("Guilds linked!").setDescription("The Guild allowed the link. Have fun!").build();
-                                                    event.getChannel().sendMessage(msgembed).queue();
-                                                    event.getJDA().getTextChannelById(rethink.getLinkChannel(requestguildid)).sendMessage(msgembed).queue();
-                                                } else if (event.getReactionEmote().getName().equals("❌")) {
-                                                    msg.delete().queue();
+                                    MessageEmbed msgembed = new MessageEditor().setDefaultSettings(MessageEditor.MessageType.INFO)
+                                            .setTitle("Guilds linked!").setDescription("The Guild allowed the link. Have fun!").build();
+                                    event.getChannel().sendMessage(msgembed).queue();
+                                    event.getJDA().getTextChannelById(rethink.getLinkChannel(requestguildid)).sendMessage(msgembed).queue();
+                                } else if (event.getReactionEmote().getName().equals("❌")) {
+                                    msg.delete().queue();
 
-                                                    MessageEmbed msgembed = new MessageEditor().setDefaultSettings(MessageEditor.MessageType.INFO)
-                                                            .setTitle("Guild linking denied.").setDescription("The Guild denied the link. :(").build();
-                                                    event.getChannel().sendMessage(msgembed).queue();
-                                                    event.getJDA().getTextChannelById(rethink.getLinkChannel(requestguildid)).sendMessage(msgembed).queue();
-                                                }
-                                            }
-                                        }
-                                    }
+                                    MessageEmbed msgembed = new MessageEditor().setDefaultSettings(MessageEditor.MessageType.INFO)
+                                            .setTitle("Guild linking denied.").setDescription("The Guild denied the linking. :(").build();
+                                    event.getChannel().sendMessage(msgembed).queue();
+                                    event.getJDA().getTextChannelById(rethink.getLinkChannel(requestguildid)).sendMessage(msgembed).queue();
                                 }
                             }
                         }
+                    } else {
+                        event.getTextChannel().sendMessage(new MessageEditor().setDefaultSettings(MessageEditor.MessageType.NO_PERMISSION).build()).queue();
                     }
                 }
         );
